+++ stories/buttons/placeholder-button.stories.js
@@ -1,3 +1,4 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { action } from '@storybook/addon-actions'
 
 import { PlaceholderButton, UiKitInitializer } from '../../src/components/index.ts'
@@ -11,6 +12,9 @@ export const NormalState = args => {
   return (
     <UiKitInitializer>
       <PlaceholderButton {...args} onClick={action('clicked')} />
+      <section style={{ display: 'flex', flexDirection: 'column', padding: '5px 0' }}>
+        <PlaceholderButton {...args} onClick={action('clicked')} />
+      </section>
     </UiKitInitializer>
   )
 }
@@ -18,4 +22,24 @@ export const NormalState = args => {
 NormalState.args = {
   children: 'Add an element',
   isDisabled: false,
+}
+
+export const WithJSXChildren = args => {
+  return (
+    <UiKitInitializer>
+      <PlaceholderButton {...args} onClick={action('clicked')} />
+      <section style={{ display: 'flex', flexDirection: 'column', padding: '5px 0' }}>
+        <PlaceholderButton {...args} onClick={action('clicked')} />
+      </section>
+    </UiKitInitializer>
+  )
+}
+
+WithJSXChildren.args = {
+  children: (
+    <>
+      <FontAwesomeIcon icon="align-left" />
+      &nbsp;Add text
+    </>
+  ),
 }