+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add an `iconPosition` attribute to allow icon position customisation
+- Add an `textPosition` attribute to allow text position customisation
+
 ## [v1.6.0] - 2021-08-06
 
 ### Addded